Fix Expanded Desktop menu
This is a squash of the following commits:

Settings: Fix expandable desktop
* Currently it does not show the per-app-setting
* Fix: When the list was rebuilt, actually display it instead of
  retriggering a rebuild...

Change-Id: Ie2db8795d16aff6bf2069016b51be28aae1ac664
Author: BadDaemon

ExpandedDesktop: Don't re-populate the app list in rebuild()
The rebuild of application entries is now done asynchronously,
so we can never expect mSession.rebuild() to return anything but null.

onRebuildComplete() correctly notifies when the rebuild is complete,
so we should only re-populate the application list there.

Change-Id: Icee8f287cac9b2b364336290a12eafe9c7800d4d
Author: zwliew