Settings: Add an option to force pre-O apps to use full screen aspect ratio [3/3]
When an app target pre-O releases, the default max aspect ratio
is 1.86:1 which leads to ugly black areas on devices that have
screens with higher aspect ratio (for example Galaxy S8/S9).

This change adds an option to allow users to change aspect ratio
for pre-O apps to full screen aspect ratio.

[akhil: adapted from https://github.com/LineageOS/android_packages_apps_LineageParts/commit/1b921f454a4d4c30317b72958251e83ff3792d81]

Change-Id: I20eeabf140c9d278d21d68675aed856d45b2f78a
Signed-off-by: Akhil Narang <akhilnarang.1999@gmail.com>