Gestures: Hide/show our different Tap-to-wake options conditionally
This is an extension of the following custom Tap-to-wake option:
https://github.com/AquariOS/packages_apps_Settings/commit/acdee90932f74ea80f8cbbad859397d1309c446f

While the custom version is being utilizied by most pixel devices,
it doesn't work for coral & flame - yet, it is still showing.
This change allows us to hide this on coral & flame without
having to add yet another device side bool.

See vendor_support for supporting method.

Change-Id: Ie15bd562bce4f37829c2e27805b9028d8a2ac512