Allow virtually any password length
The default maximum password length in AOSP is pretty low with only 16 characters. Many people complain about this. There is no downside in increasing it: People who don't want a long password can still use a short one. So I propose increasing it to 160 characters for now! I think this change is all that is needed to do so, but have not compiled and testet with this change.